backend/pyproject.toml
[tool.poetry]
name = "django-graphql-react-admin-backend"
version = "0.1.0"
description = ""
authors = ["ross <rossmere67@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
django-extensions = "^3.1.5"
psycopg2-binary = "^2.9.3"
strawberry-graphql = "^0.117.1"
strawberry-graphql-django = "^0.3.1"
strawberry-django-plus = "^1.17.0"
django-cors-headers = "^3.13.0"
django-cte = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
